refactor(pipeline): move pipeline specs to relevant renderers
